Updated app resources and version details
This commit includes several updates to the project. The reference to the Android resource `Logut.png` was corrected to `Logout.png` in `Project-K.Android.csproj`, and a new resource `food.png` was added. The application's version code and name were updated from `3` and `3.0.0` to `6` and `6.0.0` in `AndroidManifest.xml`. Resource identifiers in `Resource.designer.cs` were updated, including the correction of `Logut` to `Logout` and the change of `ic_arrow_down_24dp` to `food`. The icons for `JidelnaPage` and `LoginPage` ShellContent in `AppShell.xaml` were updated to `food.png` and `Logout.png` respectively. Lastly, two new PNG files, `food.png` and `Logout.png`, were added to the project.