Refactor HttpClient and UI elements, add ViewModel and refresh methods
In GetRozvrh.cs, the HttpClient object is now instantiated within the RefreshRozvrh method, and the ObservableCollection<Cell> Rozvrh initialization has been moved outside the static constructor. JidelnaPage.xaml has been updated with a new RefreshView and several properties of UI elements have been added or modified. A ViewModel object has been added to JidelnaPage.xaml.cs and set as the BindingContext, with the Command property of JidelnaRefresh now set to a new Command. In RozvrhPage.xaml.cs, the RefreshRozvrh method is now called in the constructor, and the ItemsSource property of RozvrhCollectionView is updated after refreshing the data.